+++ src/entry/dependenciesAny/dependenciesUnitFunction.generated.js
@@ -4,10 +4,10 @@
  */
 import { UnitDependencies } from './dependenciesUnitClass.generated'
 import { typedDependencies } from './dependenciesTyped.generated'
-import { createUnit } from '../../factoriesAny.js'
+import { createUnitFunction } from '../../factoriesAny.js'
 
 export const unitDependencies = {
   UnitDependencies,
   typedDependencies,
-  createUnit
+  createUnitFunction
 }