+++ src/entry/mainNumber.js
@@ -1,5 +1,10 @@
+// configuration
+export { config } from './configReadonly'
+
 // functions and constants
-export * from './functionsNumber'
+export * from './pureFunctionsNumber.generated'
+export * from './impureFunctionsNumber.generated'
+export * from './typeChecks'
 
 // error classes
 export { IndexError } from '../error/IndexError'