+++ test-dist/lib.test.js
@@ -1,15 +1,22 @@
+// Only use native node.js API's here, this file is not transpiled!
 const assert = require('assert')
 const version = require('../package.json').version
 
 describe('lib', function () {
-  it('should load lib/index.js', function () {
-    const math = require('..')
+  it('should load lib/mainAny.js', function () {
+    const math = require('../lib/entry/mainAny')
+
+    assert.strictEqual(math.add(2, 3), 5)
+  })
+
+  it('should load lib/mainNumber.js', function () {
+    const math = require('../lib/entry/mainNumber')
 
     assert.strictEqual(math.add(2, 3), 5)
   })
 
   it('should have the correct version number', function () {
-    const math = require('..')
+    const math = require('../lib/entry/mainAny')
 
     assert.strictEqual(math.version, version)
   })