+++ test/entry/bundleAny.test.js
@@ -1,7 +1,7 @@
 import assert from 'assert'
-import math from '../src/entry/bundleAny'
+import math from '../../src/entry/bundleAny'
 
-describe('index', function () {
+describe('bundleAny', function () {
   it('should get a default instance of mathjs', function () {
     assert.strictEqual(typeof math, 'object')
     assert.deepStrictEqual(math.config(), {