+++ tools/validateEmbeddedDocs.js
@@ -2,25 +2,25 @@
  * Validate whether all functions in math.js are documented in math.docs
  * we use the minified bundle to also check whether that bundle is valid.
  */
-const gutil = require('gulp-util')
+const log = require('fancy-log')
 const math = require('../dist/math.min.js')
 let prop
 
 // names to ignore
 const ignore = [
   // functions not supported or relevant for the parser:
-  'create', 'core', 'typed', 'config',
-  'on', 'off', 'emit', 'once',
-  'compile', 'parse', 'parser',
-  'chain', 'print', 'uninitialized',
+  'chain', 'print',
+  'compile', 'parse', 'parser', // TODO: add embedded docs for compile, parse, and parser?
+  'reviver', // TODO: add embedded docs for reviver?
+  'apply', // FIXME: apply is not supported right now because of security concerns
   'eye', 'var', 'eval', 'typeof', // deprecated functions
   'addScalar', 'divideScalar', 'multiplyScalar', 'equalScalar'
 ]
 
 // test whether all functions are documented
 let undocumentedCount = 0
-for (prop in math) {
-  if (math.hasOwnProperty(prop)) {
+for (prop in math.expression.mathWithTransform) {
+  if (math.expression.mathWithTransform.hasOwnProperty(prop)) {
     const obj = math[prop]
     if (math['typeOf'](obj) !== 'Object') {
       try {
@@ -29,7 +29,7 @@ for (prop in math) {
         }
       } catch (err) {
         undocumentedCount++
-        gutil.log('WARNING: Function ' + prop + ' is undocumented')
+        log('WARNING: Function ' + prop + ' is undocumented')
       }
     }
   }
@@ -41,7 +41,7 @@ const docs = math.docs
 for (prop in docs) {
   if (docs.hasOwnProperty(prop)) {
     if (math[prop] === undefined) {
-      gutil.log('WARNING: Documentation for a non-existing function "' + prop + '"')
+      log('WARNING: Documentation for a non-existing function "' + prop + '"')
       nonExistingCount++
     }
   }
@@ -49,7 +49,7 @@ for (prop in docs) {
 
 // done. Output results
 if (undocumentedCount === 0 && nonExistingCount === 0) {
-  gutil.log('Validation successful: all functions are documented.')
+  log('Validation successful: all functions are documented.')
 } else {
-  gutil.log('Validation failed: not all functions are documented.')
+  log.error('Validation failed: not all functions are documented.')
 }